Whoop Whoop, base stepper code fully functional!
That is AWESOME, as now those dang steppers can be used perfectly!
Also, the base frame code can easily be extended by overwriting the basic classes, or the stepper itself can be tested and moved about with the implemented function!

+++ Code/LCD.cpp
@@ -167,8 +167,8 @@
 
 	//Write out a number onto the screen, starting at "start" and displaying "len" characters (in base 10)
 	void LCD::writeNum(uint16_t n, uint8_t start, uint8_t len) {
-		uint16_t x=1;
-		uint16_t oN=0;
+		uint32_t x=1;
+		uint32_t oN=0;
 		for(uint8_t i=1; i < len; i++) {
 			x *= 10;
 		}
